pyproject.toml
[project]
name = "agentify-example-osworld"
version = "0.1.0"
description = "Example code for agentifying OSWorld using A2A and MCP standards."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "earthshaker>=0.1.12",
    "httpx>=0.27.0",
    "pydantic-settings>=2.11.0",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
    "openai>=1.50.0",
]
